Add a few bits of clean up
