+++ package.json
@@ -95,10 +95,12 @@
     "semantic-release-stop-before-publish": "^1.0.0",
     "semantic-release-vsce": "^5.5.9",
     "typescript": "^5.0.4",
-    "@types/node-fetch": "^2.6.11"
+    "@types/node-fetch": "^2.6.11",
+    "@types/xml2js": "^0.4.14"
   },
   "dependencies": {
-    "node-fetch": "3.3.2"
+    "node-fetch": "3.3.2",
+    "xml2js": "^0.6.2"
   },
   "repository": {
     "type": "git",
